+++ packages/network-enablement-controller/CHANGELOG.md
@@ -29,6 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Clean up Slip44Service ([#7626](https://github.com/MetaMask/core/pull/7626))
 - Add missing MegaETH to POPULARE_NETWORKS list ([#7286](https://github.com/MetaMask/core/pull/7286))
 
 ## [4.0.0]